adjusting log access and adding updates

+++ src/main/java/br/com/jdbcmysql/DAO/LogAccessDAO.java
@@ -2,11 +2,7 @@ package br.com.jdbcmysql.DAO;
 
 import br.com.jdbcmysql.model.User;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.Statement;
-
 public interface LogAccessDAO {
 
-    long thereWereChanges(User u);
+    int thereWereChanges(User u);
 }
